JS: factor out pretext-dialog from pretext_add_on
